Refactor vertex attributes into own module
